+++ AlirezaEShop/Models/Item.cs
@@ -3,7 +3,7 @@
     public class Item
     {
         public int Id { get; set; }
-        public decimal price { get; set; }
+        public Int64 price { get; set; }
         public int quantityInStock { get; set; }
 
         public Product product { get; set; }
